test(schemavalidate,factory): #314 format inertness tripwire; extract the factory mock registry
format is annotation-only in github.com/google/jsonschema-go, which has no
assertion switch. Create is enforcing with no escape hatch, so a dependency bump
that started asserting would begin rejecting live uuid, email and date-time
values on the write path. The spec called for this tripwire and it was missing.
A second test keeps it load-bearing by proving shipped schemas actually declare
formats.

factory_test.go grew in this PR and was 985 lines against a 500-line cap. Its
mock SchemaRegistry and the schema body it serves move to their own file — the
contained refactor of the immediate area, not a split of the whole file.
